Popravek pri sprejemanju decimalnih števil (cena), dodan decimal model binder

+++ EventiqonWebApp/EventiqonWebApp/Models/OrganizerModels.cs
@@ -19,7 +19,7 @@ namespace EventiqonWebApp.Models
         public string geolocation_lon { get; set; }
         public List<PostavkaVhodniPodatki> seznamPostavk { get; set; }
         public int steviloMestSpinner { get; set; }
-        public double cenaVstopnice { get; set; }
+        public decimal cenaVstopnice { get; set; } 
     }
     public class PostavkaVhodniPodatki
     {
